# REGR: Performance of DataFrame reduction ops with axis=1