- [ ] closes #xxxx (Replace xxxx with the GitHub issue number)
- [ ] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [ ] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [ ] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [ ] Added an entry in the latest `doc/source/whatsnew/vX.X.X.rst` file if fixing a bug or adding a new feature.

Siphoning off the largest impact from #52083.

```
shape = 250_000, 100
mask = pd.DataFrame(np.random.randint(0, 1, size=shape), dtype=pd.BooleanDtype())
%timeit mask.transpose()
# 4.28 s ± 12.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <-- main
# 1.12 s ± 19.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <-- PR
```