@jbrockmendel 2023-04-16
I've been thinking about this and #52083.  This is a workaround for lack of 2D support in MaskedArray (not unlike #52261).  Maybe its time to revisit actually supporting 2D in MaskedArray?