@rhshadrach 2023-04-18
> I've been thinking about this and #52083. This is a workaround for lack of 2D support in MaskedArray (not unlike #52261). Maybe its time to revisit actually supporting 2D in MaskedArray?

Agreed this is a workaround, no opposition to 2D for masked dtypes.